[Rename] SearchDestinationServiceTypeView으로 네이밍 변경
